change tracker callback in uploader to accept int insted of float

+++ backend/engine/uploader/Dailymotion/uploader.py
@@ -46,13 +46,13 @@ class DailymotionUploader(Uploader):
         #     'Content-Type': 'application/x-www-form-urlencoded'
         # }
 
-    def upload(self, file_path: str, progress_tracker: Callable[[float], None] = None) -> str:
+    def upload(self, file_path: str, progress_tracker: Callable[[int], None] = None) -> str:
         upload_url = self.__upload_video(file_path, progress_tracker)
         publish_url = self.__publish_uploaded_video(upload_url)
         return publish_url
 
     # Sending the video file to the upload url obtained in the previous function
-    def __upload_video(self, file_path, tracker: Callable[[float], None]):
+    def __upload_video(self, file_path, tracker: Callable[[int], None]):
 
         video_size = os.stat(file_path).st_size
         chunk_amount = video_size / UPLOAD_VIDEO_CHUNK_SIZE
